backend/cloud_functions/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.11.*